Never sort port lists during rendering
Currently, the ports are sorted to ensure interface vlans are printed
before physical interfaces. This sorting works on the port name and ends
up returning ports like 1, 10, 11, 12, ... 2, 3, etc.

This commit removes the sorting and uses proper switch_configuration
methods to obtain physical interfaces vs interface vlans. This ensures
the order specified in the configuration is kept as-is.

Closes #88