[cisco] Wrap port names on show vlan
Currently, port names were written all on one line, as if the software
connecting had defined a very long terminal. This is not always the
case.

Fake-switch should expose possible cases, such as port names wrapping on
other lines.

This commit adds this (missing) feature to the show vlan command.

This is a follow up of #90.

Partial-bug: #97