Refactor the global reactor
It's about time this thing gets a little sexier.

The test switches are now in a normalized structure, less clunky than
instantiating all the stuff, and they use the global mapping so that no
more switch is forgotten in that list.

All the tests using the test switches now have a member defining which
switch it's using and the base test classes can create the right clients
according to the configuration. This allowed the introduction of random
port numbers.

To achieve this, a contract change was required:
- The juniper switches don't create their own ports anymore, it will be
the user's responsibility for this, this is a contract change that will
trigger a major version update