Add basic vrrp groups support for juniper mx
Supports adding a VIP and a priority, more fields to come
Since VRRPs can technically be attached to anything, it was easier
for rendering to keep a record of which IP it belongs to.
Seemed easier than to try and reverse engineer from the vip to which
subnet it would belong.