Add support for identifiers on filtering
When an element has a 1 level node with a text (such as a "name") we
consider it an "identifier" if there's another tree beside it.

So any leaves under that tree should be prefixed with the path
containing the identifier.

Example:
if you look for a unit under an interface, the xpath generator would
return you

/interface/name[text()="if"]..
AND
/interface/unit/name[text()="if"]..

Which means that if the second xpath would not yield any valid endpoints
in the element filtering, only the first xpath would be consider and
return all units instead of an empty list.

Now the xpath generated looks like
/interface/name[text()="if"]../unit/name[text()="if"]..

Which means that we look for a particular child inside a given parent.