Fix deployment error (#125)
When tagging and deploying on pypi, the python: directive indicates
which environment to deploy in. The current matrix has 2 jobs
matching python 2.7, the 2.7 tests and the docker integrated tests so
when we tag it tries to deploy in both and one ends up failing because
the package already exist on pypi.  It used to not bother travis but
that might have been a bug.

Switching to python 3.5 will make sure only 1 job tries to deploy.