Add a display driver mechanism
Having a special display driver, commands are able to format their
objects and send them to the display that will then display it.

Apart from extracting display information from the word processor, this
will enable the eapi to inject a JSON Display that will pass the objects
to the api rather than render them.