Use a unique display object (#139)
Having one for each processor was indeed simplifying how to pass the
output channel (the processor itself) on instantiation but it causes a
problem in the EAPI since that service holds a reference to the base
processor only and the base processor is the one dispatching down its
tree so it's hard to know which processor processed the command to get
that processor's display to get the output object.

Having a unique display and passing the output channel on each call will
solve this as we don't care who processed it.

We cannot simply pass the terminal controller to the display since there
is a piping mechanism in the .write that would need to be copied. A
refactor could turn that into a mixin, or the terminal controller could
controller the piping.